LayoutTests/webrtc/audio-peer-connection-webaudio.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>Testing local audio capture playback causes "playing" event to fire</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src ="routines.js"></script>
    <script>
    promise_test((test) => {
        if (window.testRunner)
            testRunner.setUserMediaPermission(true);

        return navigator.mediaDevices.getUserMedia({audio: true}).then((stream) => {
            if (window.internals)
                internals.useMockRTCPeerConnectionFactory("TwoRealPeerConnections");
            return new Promise((resolve, reject) => {
                createConnections((firstConnection) => {
                    firstConnection.addStream(stream);
                }, (secondConnection) => {
                    secondConnection.onaddstream = (streamEvent) => { resolve(streamEvent.stream); };
                });
                setTimeout(() => reject("Test timed out"), 5000);
            });
        }).then((remoteStream) => {
            return analyseAudio(remoteStream, 1000);
        }).then((results) => {
            assert_true(results.heardHum, "heard hum");
            assert_true(results.heardBip, "heard bip");
            assert_true(results.heardBop, "heard bop");
        });
    }, "Basic audio playback through a peer connection");
    </script>
</head>
<body>
</body>
</html>
